* main: (52 commits)
  remove obsolete image/description from luke/README.md
  Upgrade to forbiddenapis 3.3 (#768)
  LUCENE-10393: Unify binary dictionary and dictionary writer in kuromoji and nori (#740)
  LUCENE-9651 Update benchmark module docs (#759)
  LUCENE-10458: BoundedDocSetIdIterator may supply error count in Weigth#count(LeafReaderContext) when missingValue enables (#736)
  LUCENE-10481: FacetsCollector will not request scores if it does not use them (#760)
  LUCENE-10477: mention 'call multiple times' in Query.rewrite javadoc (#758)
  Add back-compat indices for 9.1.0.
  Synchronize CHANGES.
  LUCENE-10464, LUCENE-10477: WeightedSpanTermExtractor.extractWeightedSpanTerms to rewrite sufficiently (#737)
  Add version 9.1.0.
  DOAP changes for release 9.1.0
  LUCENE-10422: Make errorprone happy
  LUCENE-10478: mark Test4GBStoredFields as @Monster (#757)
  LUCENE-10422: Read-only monitor implementation (#679)
  LUCENE-10473: Make tests a bit faster when running nightly. (#754)
  LUCENE-9905: Fix check in TestPerFieldKnnVectorsFormat#testMergeUsesNewFormat
  LUCENE-9614: Fix rare TestKnnVectorQuery failures
  LUCENE-10472: Fix TestMatchAllDocsQuery#testEarlyTermination (#753)
  LUCENE-10418: Move CHANGES to the correct section.
  ...

# Conflicts:
#	lucene/sandbox/src/java/org/apache/lucene/sandbox/search/IndexSortSortedNumericDocValuesRangeQuery.java
#	lucene/sandbox/src/test/org/apache/lucene/sandbox/search/TestIndexSortSortedNumericDocValuesRangeQuery.java

+++ lucene/test-framework/src/java/org/apache/lucene/tests/index/BaseKnnVectorsFormatTestCase.java
@@ -594,7 +594,7 @@ public abstract class BaseKnnVectorsFormatTestCase extends BaseIndexFileFormatTe
     int numDocs = atLeast(1000);
     int numFields = TestUtil.nextInt(random(), 1, 10);
     int[] fieldDocCounts = new int[numFields];
-    float[] fieldTotals = new float[numFields];
+    double[] fieldTotals = new double[numFields];
     int[] fieldDims = new int[numFields];
     VectorSimilarityFunction[] fieldSearchStrategies = new VectorSimilarityFunction[numFields];
     for (int i = 0; i < numFields; i++) {
@@ -622,7 +622,7 @@ public abstract class BaseKnnVectorsFormatTestCase extends BaseIndexFileFormatTe
       try (IndexReader r = w.getReader()) {
         for (int field = 0; field < numFields; field++) {
           int docCount = 0;
-          float checksum = 0;
+          double checksum = 0;
           String fieldName = "int" + field;
           for (LeafReaderContext ctx : r.leaves()) {
             VectorValues vectors = ctx.reader().getVectorValues(fieldName);
@@ -1095,7 +1095,7 @@ public abstract class BaseKnnVectorsFormatTestCase extends BaseIndexFileFormatTe
         VectorSimilarityFunction.values()[
             random().nextInt(VectorSimilarityFunction.values().length)];
 
-    float fieldValuesCheckSum = 0f;
+    double fieldValuesCheckSum = 0;
     int fieldDocCount = 0;
     long fieldSumDocIDs = 0;
 
@@ -1120,7 +1120,7 @@ public abstract class BaseKnnVectorsFormatTestCase extends BaseIndexFileFormatTe
       }
 
       try (IndexReader r = w.getReader()) {
-        float checksum = 0;
+        double checksum = 0;
         int docCount = 0;
         long sumDocIds = 0;
         for (LeafReaderContext ctx : r.leaves()) {
